backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
groq>=0.11.0
python-multipart==0.0.17
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
PyPDF2==3.0.1
openpyxl==3.1.5
python-docx==1.1.2
redis==5.2.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0
